requirements.txt
pyaes==1.6.1
pyasn1==0.4.8
PySocks==1.7.1
pytz==2021.1
rsa==4.7.2
Telethon==1.21.1
